rolex: config.fs: init: Add AID_VENDOR_QRTR for qrtr-ns
Add the AID_VENDOR_QRTR AID in config.fs. Change qrtr-ns to use this
AID for system and group. This AID is reserved for applications that
perform control operations on the QRTR socket family.

The QRTR node ID is hardcoded to 1 within the kernel. Do not configure
the node from userspace. This allows qrtr-ns to drop NET_ADMIN
capabilities.

Signed-off-by: SunnyRaj84348 <sunnyraj7461@gmail.com>

Change-Id: Iab00f2e3e8203319abb2c6bf84172cf239056fa8